+++ react/test.tsx
@@ -1,6 +1,7 @@
 import { createRoot } from 'react-dom/client';
 import React, { useEffect, useRef } from 'react';
 import { Comments, CommentCount, NewsletterForm, Memberships, GatedContent }  from './src';
+import { CommentCounts } from '@hyvor/hyvor-talk-base';
 
 const root = createRoot(document.getElementById('root'));
 root.render(<App />);
@@ -19,13 +20,18 @@ function App() {
                 membershipsRef.current?.element(),
             );
         },)
-
     }, [])
 
     return <div>
         <Comments
             website-id={14}
             ref={commentsRef}
+            on={{
+                "comment:published": (comment) => {
+                    comment.id;
+                },
+                "auth:login:clicked": (d) => console.log(d)
+            }}
         />
         <CommentCount
             page-id="test"